pyproject.toml
[project]
name = "yabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appdirs>=1.4.4",
    "langgraph>=0.2.0",
    "matrix-nio[e2e]>=0.25.2",
    "openai>=2.15.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]

[project.scripts]
yabot = "yabot.app:run"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["yabot", "yabot.*", "matrix_bot", "matrix_bot.*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
